improved glue interaction, the glue now removes rigidbodys if collided with a child element, implemented a better suited evaluation architecture
